+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add neg function to invert sign of numeric values
+- Add cast functions for numeric values
+
+
 ## [0.1.3] - 2022-11-26
 
 ### Added
@@ -51,7 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement a `RuleSet` as an entry point for executing a set of rules on a piece of data 
 - async `UserFunction` support
 
-[unreleased]: https://github.com/mendelt/reval/compare/v0.1.2...HEAD
+[unreleased]: https://github.com/mendelt/reval/compare/v0.1.3...HEAD
 [0.1.3]: https://github.com/mendelt/reval/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/mendelt/reval/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/mendelt/reval/compare/v0.1.0...v0.1.1